MRFP/requirements.txt
sacrebleu==1.5.1
pytorch-lamb==1.0.0
transformers==4.8.2
IPython==7.25.0
git-python==1.0.3
rouge_score==0.0.4
datasets==1.9.0
scikit-learn
beautifulsoup4==4.9.3
matplotlib==3.4.2
pynvml==11.0.0
tensorboard==2.5.0
networkx
validators
spacy==2.3.5
nltk==3.5
rouge==1.0.1
wandb==0.12.1
scikit-multilearn
iterative-stratification==0.1.6
bert_score==0.3.7
pymediawiki
flair==0.9
pymediawiki
translators
tweet-preprocessor
gsitk
google-cloud-translate==2.0.1
googletrans
emoji